+++ Common/src/main/java/fuzs/puzzleslib/api/client/event/v1/renderer/FogEvents.java
@@ -25,29 +25,30 @@ public final class FogEvents {
          *
          * @param gameRenderer the game renderer instance
          * @param camera       the camera instance
-         * @param partialTicks partial ticks for this rendering operation
+         * @param partialTick  partial ticks for this rendering operation
          * @param fogMode      fog mode
          * @param fogType      fog type
          * @param fogStart     distance from the camera for the fog to start
          * @param fogEnd       distance from the camera for the fog to end
          * @param fogShape     spherical or cylindrical fog shape
          */
-        void onRenderFog(GameRenderer gameRenderer, Camera camera, float partialTicks, FogRenderer.FogMode fogMode, FogType fogType, MutableFloat fogStart, MutableFloat fogEnd, MutableValue<FogShape> fogShape);
+        void onRenderFog(GameRenderer gameRenderer, Camera camera, float partialTick, FogRenderer.FogMode fogMode, FogType fogType, MutableFloat fogStart, MutableFloat fogEnd, MutableValue<FogShape> fogShape);
     }
 
     @FunctionalInterface
     public interface ComputeColor {
 
         /**
-         * Called after the fog color is calculated from the current block overlay or biome. Allows for modifying the fog color.
+         * Called after the fog color is calculated from the current block overlay or biome. Allows for modifying the
+         * fog color.
          *
          * @param gameRenderer the game renderer instance
          * @param camera       the camera instance
-         * @param partialTicks partial ticks for this rendering operation
+         * @param partialTick  partial ticks for this rendering operation
          * @param fogRed       red color component
          * @param fogGreen     green color component
          * @param fogBlue      blue color component
          */
-        void onComputeFogColor(GameRenderer gameRenderer, Camera camera, float partialTicks, MutableFloat fogRed, MutableFloat fogGreen, MutableFloat fogBlue);
+        void onComputeFogColor(GameRenderer gameRenderer, Camera camera, float partialTick, MutableFloat fogRed, MutableFloat fogGreen, MutableFloat fogBlue);
     }
 }